First full working solution #7
