+++ client/src/games/Game.tsx
@@ -5,6 +5,11 @@ import { Controls } from "../gameLogic/controls";
 import { GameLoop } from "../gameLogic/GameLoop";
 import { useNavigate } from "react-router-dom";
 import Renderer from "../util/Renderer";
+import Spinner from "../util/Spinner";
+import {
+  connectToWebSocket,
+  waitForOpponent,
+} from "../background/worker-wrapper";
 
 const Game = () => {
   const navigate = useNavigate();
@@ -41,6 +46,8 @@ const Game = () => {
 
   const worker = useRef<Worker | null>(null);
 
+  const [gameStarted, setGameStarted] = useState<boolean>(false);
+
   const backToHome = () => {
     if (worker.current) {
       worker.current.postMessage({
@@ -54,13 +61,26 @@ const Game = () => {
 
   useEffect(() => {
     const redirectToFullRoom = () => {
+      worker.current?.terminate();
       navigate(`/fullRoom`);
     };
+
     worker.current = new Worker(
       new URL("./../background/socket-worker.ts", import.meta.url)
     );
 
-    // connectToWebSocket(roomId, worker as React.MutableRefObject<Worker>);
+    connectToWebSocket(roomId, worker as React.MutableRefObject<Worker>);
+
+    waitForOpponent(worker.current, setGameStarted, redirectToFullRoom);
+  }, [worker, roomId, navigate]);
+
+  useEffect(() => {
+    if (!gameStarted || !worker.current) return; // Early return if not ready
+
+    console.log("Opponent is there");
+    const redirectToFullRoom = () => {
+      navigate(`/fullRoom`);
+    };
 
     const backToLogin = () => {
       if (worker.current) {
@@ -73,6 +93,7 @@ const Game = () => {
       navigate("/login");
     };
 
+    // Start socket listening when game starts
     socketListen(
       setOpponentPosition,
       setLastActive,
@@ -85,9 +106,10 @@ const Game = () => {
       worker.current?.postMessage({ type: "close" });
       worker.current?.terminate();
     };
-  }, [navigate, roomId]);
+  }, [gameStarted, navigate, roomId]);
 
   useEffect(() => {
+    if (!gameStarted) return;
     const controls = new Controls(
       playerPosition,
       playerRadius,
@@ -101,9 +123,10 @@ const Game = () => {
     return () => {
       controls.removeListeners();
     };
-  }, [playerPosition]);
+  }, [gameStarted, playerPosition]);
 
   useEffect(() => {
+    if (!gameStarted) return;
     const interval = GameLoop({
       playerRadius,
       playArea,
@@ -130,19 +153,33 @@ const Game = () => {
     playerPosition.dy,
     playerPosition.x,
     playerPosition.y,
+    gameStarted,
   ]);
 
   return (
-    <Renderer
-      roomId={roomId}
-      playArea={playArea}
-      isDragging={isDragging}
-      playerPosition={playerPosition}
-      lineEnd={lineEnd}
-      playerRadius={playerRadius}
-      opponentPosition={opponentPosition}
-      loginRedirect={backToHome}
-    />
+    <>
+      {!gameStarted ? (
+        <div className="loader-container">
+          <Spinner />
+          <p>Waiting for other player</p>
+          <br />
+          <button onClick={backToHome} className="button-primary">
+            Return to login
+          </button>
+        </div>
+      ) : (
+        <Renderer
+          roomId={roomId}
+          playArea={playArea}
+          isDragging={isDragging}
+          playerPosition={playerPosition}
+          lineEnd={lineEnd}
+          playerRadius={playerRadius}
+          opponentPosition={opponentPosition}
+          loginRedirect={backToHome}
+        />
+      )}
+    </>
   );
 };
