+++ client/src/main.tsx
@@ -1,4 +1,4 @@
-import { StrictMode } from "react";
+// import { StrictMode } from "react";
 import { createRoot } from "react-dom/client";
 import "./index.css";
 import App from "./App.tsx";
@@ -9,9 +9,9 @@ const container = document.getElementById("root");
 const root = createRoot(container!);
 
 root.render(
-  <StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </StrictMode>
+  //<StrictMode>
+  <BrowserRouter>
+    <App />
+  </BrowserRouter>
+  //</StrictMode>
 );
